Bug 1633489 [wpt PR 23272] - Treat floorOrigin as mojoFromFloor, a=testonly
Automatic update from web-platform-tests
Treat floorOrigin as mojoFromFloor

Per the webxr-test-api all origins are based on the "local" space when
the "session" started; which means that all of our origins passed in are
inherently "mojoFromFoo."  The standingTransform (in need of renaming)
is, in blink and by devices, floorFromMojo. Since the tests were
directly setting floorOrigin to the standingTransform, the tests were
incorrectly treating floor origin as floorFromMojo. This change inverts
the floorOrigin before assigning it to the standingTransform, so that
Chrome begins treating the floorOrigin as mojoFromFloor. Given that
Chrome was the only ones running these particular tests, the output
values were *also* written as if the passed in floorOrigins were
floorFromMojo. In order to stick with the expectations, this inverts
the input transforms as well.

In fixing this, a bug was discovered in XRMathHelper.inverse wherein the
comatrix was not created properly.  This change also addresses that bug.

Fixed:1071630

Change-Id: I1626d80099169c4ceba0df68ad84bbf181f76647
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165881
Auto-Submit: Alexander Cooper <alcooperchromium.org>
Commit-Queue: Piotr Bialecki <bialpiochromium.org>
Reviewed-by: Piotr Bialecki <bialpiochromium.org>
Cr-Commit-Position: refs/heads/master{#763104}

--

wpt-commits: 6a2f0e6a0110fd0830fc3196680f671bb83c2747
wpt-pr: 23272

UltraBlame original commit: 01615f7b96b20133e233d44e538e29ff66b175c8