Bug 1633969 - Refresh .git-blame-ignore-revs. r=sylvestre
It was generated by mixing every commit with "ignore-this-changeset" and
everything listed in .hg-annotate-ignore-revs, in topologic order.

The format was changed to:

  ```
  # hg-sha1 - author - subject
  cinnabar-sha1 # cinnabar
  gecko-dev-sha1 # gecko-dev
  ```

The comment line corresponds to entries in .hg-annotate-ignore-revs (or
what they would look like for commits with "ignore-this-changeset").

Differential Revision: https://phabricator.services.mozilla.com/D73210

UltraBlame original commit: 02cbfa7967b274d25d098b98fd3948a71e4e4821