Bug 1634377 part 2: Fix WPT reftest align-content_stretch.html to avoid fractional pixels in spacing. r=TYLin
Before this change, this test ends up with 50px of spacing, divided 3 ways
(with one share going after each flex item). That means we should be expecting
spacing of 50px/3 = 16.666px -- but the reference case was instead expecting
arbitrarily-rounded amounts, expecting 17px of spacing in one spot and 16px in
another spot (which may or may not match actual UA behavior, particularly on a
HiDPI display where pixel-snapping might not need to round to whole CSS
pixels).

This patch just increases the size of the container so that we end up with
60px of extra space, which gives us nice round number of pixels for each share
of the space: 60px/3 = 20px of space between each flex item.

(This is consistent with the neighboring test "align-content_space-around.html"
which also uses a 210px-tall container for the same reason.)

Also: While we're at it, swap in spaces for a few stray tab characters
in some contextual lines.

Differential Revision: https://phabricator.services.mozilla.com/D73549

UltraBlame original commit: 42838759a8f61c52aa2da6d59839a757028221cb