Bug 1626029 - Don't attempt to resend mDNS packets on EPERM; r=mjf
We currently retry sending queries on all failures to write to the socket on
the assumption that write errors represent transient errors. With this change,
we treat EPERM has a permanent error and drop queries if we see it. This is
not a problem when sending answers, as we don't retry in case of errors there,
on the assumption that the other side will retry their query in the future.

We do have code to limit the number of times we attempt an mDNS query, but we
didn't hit it in this case, because it only counts sends that are successful.

Differential Revision: https://phabricator.services.mozilla.com/D73492

UltraBlame original commit: 1c17543d6342d704456776f491b0328f921371c6