Bug 1634204 - Upgrade Linux build docker images to Debian 8. r=froydnj
This gets rid for the need of a number of local packages (mostly related
to Gtk+3). One exception is that we now need a 32-bits version of the
xz-utils package, some -dev package depends on it, and that dependency
can't be fulfilled in the 32-bits image because we already have the
64-bits backport installed, which conflicts with it (we need both
32-bits and 64-bits package to be at the same version when installed).

The system binutils fails to link clang-7 for some reason, so we now use
our toolchain binutils instead, like we already do for newer versions of
clang.

The debian-packages docker image now needs an explicit installation of
git, because it's not pulled in via the recommends of some other
package.

For some reason, snapshot.debian.org doesn't contain the jessie-backports
archive at the same location as others, and only has a few snapshots of
the archive.

Differential Revision: https://phabricator.services.mozilla.com/D73784

UltraBlame original commit: 8621fde19c3154f11758d8f9e0d2a22cfff1efed