Bug 1634535 - Move ply to third_party/python r=glandium
The license used to be LGPL so the code lived in other-licenses, but it was changed to BSD eleven years ago. Let's move it over to third_party/python/ply where it belongs.

    ./mach vendor python ply==3.10

`diff -r` between the original `ply` directory and the new one only comes up with the new file `third_party/python/ply/CHANGES` which isn't relevant to the functionality of the code, so this should be a no-op all told.

Differential Revision: https://phabricator.services.mozilla.com/D73341

UltraBlame original commit: 78c07ca17076dd63902f7b5ad2461bbbaf8ff504