Bug 1635852: Include minidump_stackwalk files in taskgraph sparse-profile; r=glandium
Without this, taskgraph calculates the digest based on just the files in these
directories that are in the sparse profile (I suspect this is just the
moz.build files), and will rebuild it when those files change. This changes
ensures that all files in those directories are used to calculate the digest.

Additionally, this will cause the same digest to be generted by developers
locally, since the files *are* present in the non-sparse checkout that most
developers have.

Differential Revision: https://phabricator.services.mozilla.com/D74113

UltraBlame original commit: be3aa8f30db0246ecee9ba1ed146a5969e8fbd3d