Bug 1628073 - Make the defaults for the profiler settings nicer for mochitests; r=gerald
The defaults should serve our users. I did some profiling with mochitests
and the defaults weren't ideal. It would be nice if the profiler "just worked"
for most of the situations. It would be great to teach the environment variables
about presets, but for now this will make the profiler easier to use for our
users in the general case.

This patch adds both screenshots and updates the entries count. This causes
additional overhead, but people on low-end devices can still tweak the settings
to remove that overhead.

Differential Revision: https://phabricator.services.mozilla.com/D70091

UltraBlame original commit: 619d4f31f851fd19c4291dff08dc873b7f839d1b