Bug 1624550 - P5: Fix racy check for SessionHistory update. r=kmag
This code was relying on SessionHistory update having completed when
GlobalHistory is update. Moving GlobalHistory update to the parent process with
DocumentChannel means that it's possible for the GlobalHistory event to fire
before the SessionHistory is updated in nsDocShell.

Switch to using tab loading complete to signal when it's OK to check
SessionHistory.

Differential Revision: https://phabricator.services.mozilla.com/D72280

UltraBlame original commit: dd1197d4200ef5a4529483ff889481786de7a799