Bug 1636235 - Fix corner case SIMD bugs. r=wingo
There was a missing guard when freeing the SIMD result register; it
was freed unconditionally, but this would fail for br_table, when only
the int register is needed.

There was a missing case when materializing constants during
popStackResults, we need to handle V128 here.

In both cases the underlying cause was iffy test coverage, though in
the latter case the problem is also that we allow the use of switch
with default for enum class value sets, thus masking problems where a
case that must be covered is not covered.

Differential Revision: https://phabricator.services.mozilla.com/D74380

UltraBlame original commit: 9f3939d739c5571980439576ef0f2f7cd4f15378