Bug 1618695 - Forward local tab instance to tab descriptor r=ochameau
Depends on D73777

The filter is a special argument of TabDescriptor::getTarget that is
closely linked to local-tab. It can only be set by rootFront getTab.

However if we reuse the same tab descriptor for several targets, we need
to make sure to consistently create LocalTabTargetFronts if we are debugging
a local tab.

Usually the tab descriptor target is created via getTab and cached, but under
some circumstances, we can attempt to call getTarget on a tab descriptor from
another codepath first (eg reparent remote frame).

The localTab becomes a property of the TabDescriptorFront, and getTarget takes
no argument, which is consistent with other target descriptors.

Differential Revision: https://phabricator.services.mozilla.com/D73778

UltraBlame original commit: c2e6ab53a7ebc20c4f1e09e3ce613a13034f9b9c