Bug 1611096 - Check webprogresslistener as a fallback for non loaded documents r=rcaliman,ochameau
Depends on D62624

With the previous implementation, an uninitialized document could be returned as a root node.
Here we try to be more explicit and wait for a correct root node. However in some cases a document can remain uninitialized and will never transition to any other state.
If the document is uninitialized but is not currently loading, we should consider it as a valid root node.

Differential Revision: https://phabricator.services.mozilla.com/D62625

UltraBlame original commit: cc99591503474b92532fe016a10d5efc65947474