Bug 1636516 - When moving an element to another document, don't drop nsMappedAttributes from its sheet. r=edgar
That is, drop the reference to the sheet pointer. GetModifiableMapped
doesn't ever return anything that's on the hash. But trying to
DropMappedAttributes it will remove not that instance, but any other
instance that's already in the hash and matches it.

ThisDropSheetReference is always done in MakeMappedUnique if relevant.

On the original test-case it was about the -moz-user-modify property
that contenteditable maps, but it reproduces with any mapped attribute
like hidden, so I used that for WPT.

Differential Revision: https://phabricator.services.mozilla.com/D74498

UltraBlame original commit: 7365c2beddabe67206e9368f903e72878490d4dc