Bug 1629952 - Remove legacy addon install.cancel() support. r=aswan,mixedpuppy
This patch removes some old code for non-restartless addons. When the code was
written, it was still possible to call `cancel()` on an addon after the addon
was installed and before it had persisted in the database.

This case is not possible anymore, and calling `cancel()` on a
`STATE_INSTALLED` install object doesn't actually do much (but makes the
calling code think it actually canceled successfully, since it doesn't throw).

After this patch, calling `cancel()` on a `STATE_INSTALLED` install will throw
an exception.

There is one spot where we still keep around the install object because we
don't call `_cleanup()`, i.e. when installing an already present disabled
addon. For this case, we can just call `_cleanup()` at the end of the
enablement process.

There are quite a few tests that exploit calling `cancel()` to mean
`uninstall()`, for those actually calling `uninstall()` works fine.

This fixes a GeckoView bug where calling `cancel()` doesn't throw (so GV thinks
the cancellation went through successfully) but then we receive an addon object
from the successfull installation.

Differential Revision: https://phabricator.services.mozilla.com/D74305

UltraBlame original commit: 786017cc292de7d710c359b6da02bbb475953be4