Bug 1637330 - Flush frames a bit more in nsXULPopupManager. r=tnikkel
Chrome inserts shadow dom lazily during that event, which can make us
reconstruct the menupopup temporarily.

When lazyfc is disabled we reconstruct it immediately so this code gets
a non-null primary frame, but with it enabled we need to flush frames.

Do the same in popuphiding for consistency.

Differential Revision: https://phabricator.services.mozilla.com/D74887

UltraBlame original commit: 358fb28cf7035647a0a273be7eaabface6f42763