Bug 1628240 - Do not clear all origins that match the pattern for QuotaCleaner.deleteByHost; r=dom-workers-and-storage-reviewers,janv
We do clearStoragsForPrincipal based on the result from listOrigin, so it's fine
to not clear all stroages that match the prefix here. Also, since the passing
principals can contain origin attributes which is not allowed to be used with
aClearAll.

Differential Revision: https://phabricator.services.mozilla.com/D73417

UltraBlame original commit: bb7b37cfeea0f1597dc599af2148755eeca30420