Bug 1637140 - Upgrade hashbrown to 0.7.2. r=emilio
The upgrade of mp4parse pulled some code that previously was eliminated as
dead code in libxul. That code was unfortunately unreproducible as it
was generating a random seed at compile time. This was due to hashbrown
enabling the ahash feature that does that, but that was changed in 0.7.2.

Differential Revision: https://phabricator.services.mozilla.com/D75033

UltraBlame original commit: 571152088848cb0945a3cf62e4a3c1bc782f4e33