Bug 1637254, vendor l10n python libraries updates, r=stas
This updates compare-locales to 8.0. The major version bump is due
to the changes to the json output.
This also updates fluent.syntax to 0.17.
Vendor in fluent.migrate 0.9. This is the first formal vendoring,
the version we had in-tree wasn't released on pypi before.

Differential Revision: https://phabricator.services.mozilla.com/D75127

UltraBlame original commit: 07f12885bcc5b96f16b7be5199fe68355355cca1