Bug 1636886 - Only download conditioned-profiles once in Raptor. r=tarek,perftest-reviewers,Bebe
This patch makes it so that we only download and extract the conditioned-profile once and have tests use copies of that original copy. It also adds some debugging logs to the condprof `download_file` function to debug an issue where we seem to hit a cache-miss when we have already downloaded the artifact.

Differential Revision: https://phabricator.services.mozilla.com/D75176

UltraBlame original commit: 4dd20d7e4ed3864c31e5182d458b4d66012ca228