Bug 1593226 - [remote] Create a single context observer per content session. r=remote-protocol-reviewers,maja_zf
Domains would have created their own instance of the context observer,
which results in duplicated event listeners and observer notifications
to be registered.

This is still not ideal for the observer notifications, which should be
registered only once, but still an improvement for now. Bug 1635568 will
finally fix that.

Differential Revision: https://phabricator.services.mozilla.com/D74632

UltraBlame original commit: 41a0320ca96229e3096b37cb95eda563eaaa64c5