Bug 1638114 - Do not keep EGLSurfaceOverride references to destroyed Android Shared Surfaces. r=jgilbert,snorp,rbarker
In WebVR we destroy the mVRScreen after presentation is exited by calling WebGLContext::ClearVRFrame(). We found that when having multiple windows with active WebGL contexts in Firefox Reality, the WebGL context dies a lot, and you have to reload the page to recover from it. The problem is that when the GLScreenBuffer is destroyed the EGLContext still keeps a reference to the destroyed surface, set by gl->SetEGLSurfaceOverride(surface), which can break the future eglMakeCurrent calls.

Differential Revision: https://phabricator.services.mozilla.com/D75413

UltraBlame original commit: f12050b298470f8bc99e19c1976a1eed3cf63f21