Bug 1637667 - Switch into gray marking stats phase when marking gray things r=sfink
This makes GCMaker::markUntilBudgetExhaused switch into the appropriate stats phase when marking gray things. We have to pass a parameter to make this conditional because we don't do this type of report off main thread and this happens when we mark on a background thread during sweeping. I rearranged the code so that we don't ever encounter gray things in that case by doing the marking on the main thread in cases where we would have joined the background task immediately anyway. We still need to skip reporting for marking done by processMarkQueue() but that's a debugging feature so this doesn't matter.

I had to move the SWEEP_MARK_GRAY_WEAK phase under SWEEP_MARK_WEAK but that does make some sense.

Differential Revision: https://phabricator.services.mozilla.com/D75270

UltraBlame original commit: 5e5a36b8ce9e6fa8bc8c1fe6d0b36a8fc98d18b7