Bug 1421659 - Keep console output pinned to bottom when resizing. r=bomsy.
When resizing the output, if the height was modified, the scrollbar position
was updated and could lose its pinned-to-bottom state.
This patch adds a resizeObserver to the console output node, where we
scroll the output to the bottom if it's needed.
In order to be able to do that, some changes were done in how we were
detecting that the output was pinned to the bottom.
We now have an intersection observer on the last message element,
which flips a flag when the element isn't intersecting the viewport
anymore.
Some test cases are added to ensure the console has the expected
behavior in different situations.

Depends on D74708

Differential Revision: https://phabricator.services.mozilla.com/D75297

UltraBlame original commit: 8c49d2500a1690d86774659d82012ac29d4b5435