Bug 1631230 - Remove crop property. r=Honza,bradwerth,bomsy
In network analysis in Netmonitor, the headers are being cropped, even when there is space available. This is due to the `cropLimit` property. In this revision, the `cropLimit` property is removed so that the headers take up all available space.

Differential Revision: https://phabricator.services.mozilla.com/D75688

UltraBlame original commit: ffd2e6786655453f805a281111bcc184dc47ad82