Bug 1638001 - Add test name from metadata to perfherder data. r=acreskey,mozperftest-reviewers,tarek
This patch adds a `test_name` field to the test metadata that can be used to change the name of the test when output in the perfherder format. It also implements this for the app-link tests.

Differential Revision: https://phabricator.services.mozilla.com/D75544

UltraBlame original commit: d34c604f3e8b8438f87b78c454e75df320b7c575