Bug 1638821 - Make the decision to whether paint focus rings depend on the theme, not the platform. r=spohl
The current code assumes that nsNativeTheme is in use which breaks with
the non-native theme.

Instead of hackily remove the FOCUS bit, make the theme check for the
FOCUSRING bit, which is the right thing to check for anyway.

Differential Revision: https://phabricator.services.mozilla.com/D75782

UltraBlame original commit: b8259db1782b52d2ff53e131b3e531a90927852c