Bug 1559861: WR - Scale picture tasks based on their surface to parent transform scale factors r=gw,Bert
This patch adjusts the device pixel scale for raster root surfaces so that the surface's content is not
pixelated/blurred when it is composited into its parent surface where a transform scales up the surface.

This only fixes the referenced bug when WebRender is used.

Differential Revision: https://phabricator.services.mozilla.com/D75236

UltraBlame original commit: 41acfa78dae455503214f4beb23d75ac1995332b