Bug 1639524 - Fix another specificity test for :is(). r=jwatt
Same as in D75831, the test was wrong, the selectors assumed the
combinator applied to the wrong thing, and thus weren't matching.

The test fails the same way as in Nightly in WebKit, and passes with
these changes as expected.

Differential Revision: https://phabricator.services.mozilla.com/D76126

UltraBlame original commit: 5415da14ec9a2f4749e8f405d6111a6f40e8138f