Bug 1633340 - Part 6: Use the 'http' scheme for TEST_ANOTHER_3RD_PARTY_DOMAIN in ETP test for Fission. r=baku
Due to the fact the the same domain with different schemes would be
consider as third-party in Fission. So, we have to use the same scheme
when doing the test in Fission. This patch change
TEST_ANOTHER_3RD_PARTY_DOMAIN to http scheme in Fission mode. In
non-Fission mode, it will still use https scheme.

Differential Revision: https://phabricator.services.mozilla.com/D75979

UltraBlame original commit: ec7583769b9cb9127a2753cafb1172908e066ab9