Bug 1639627: [css-flex] Make flex-container percent 'row-gap' values resolve to 0, when they're resolved against an indefinite block-size. r=TYLin
This implements a CSSWG resolution in this issue:
https://github.com/w3c/csswg-drafts/issues/5081

This also effectively backs out our changes from
https://bugzilla.mozilla.org/show_bug.cgi?id=1473047
which is where we implemented our current behavior.

Note that there's no change for column-gap behavior, because this resolution is
only about block-axis sizes, which is always the "row"/row-gap axis (regardless
of the flex container's orientation / writing-mode).

I'm fixing the reftest that we have on this behavior to expect that all of its
row-gaps (which are all in the affected category) resolve to 0; and I'm
adjusting our expectations for its WPT reimport to expect failure until the
test changes have made the roundtrip back into our WPT reimport.

(Thanks to David Grogan for the new informational <link> & <meta>-tags in the
testcase.)

Differential Revision: https://phabricator.services.mozilla.com/D76186

UltraBlame original commit: 5ed4c347cc66266d13ce2289ed55df9101affa70