Bug 1639756 - Cleanup html.css using :is(). r=jwatt
There's some lists code that I could technically clean up a bit more, but I
erred in the side of making the selectors as fast as possible with our current
infrastructure.

For example, this selector list:

    :is(ul, ol, dir, menu, dl) ul,
    :is(ul, ol, dir, menu, dl) ol,
    :is(ul, ol, dir, menu, dl) dir,
    :is(ul, ol, dir, menu, dl) menu,
    :is(ul, ol, dir, menu, dl) dl

Could be reduced to:

    :is(ul, ol, dir, menu, dl) :is(ul, ol, dir, menu, dl)

But that means that for `dl` elements we'll selector-match all the selectors
inside the :is() instead of just `dl`. Maybe it doesn't matter compared with
the work of going up all the parent chain, but I erred in the side of caution
for most of these.

Differential Revision: https://phabricator.services.mozilla.com/D76265

UltraBlame original commit: ff407bdf357e75aa51b10623a2d249a4dc57e863