Bug 1640814 - Add a targetFront property on each resource. r=ochameau.
This will allow resource consumers to have a direct access to the target front
a resource was issued from.
At the moment we only use it in the console, where the targetFront is placed
on the console messages.

Differential Revision: https://phabricator.services.mozilla.com/D76790

UltraBlame original commit: b8e6c8e4bc69292c89268f205e744aed8f78f131