Bug 1640259 - Add EPERM(Operation not permitted) as an expected error in isPidAlive() check. r=ahal
We could hit this error if we open and close too many processes during
the test. This could happen when Fission is enabled. The pid number
could roll back to 1 if the pid is counted over than 32767 in linux. If
this happened, we might check the pid which is opened by others. For
example, we open and close the pid 1000 and after that the pid rolls back
to 1  and the pid 1000 has been used by other process for any reasons.
If this happens, we would hit this error during killing zombie processes
in the end of the test.

To resolve this, we can add EPERM as an expected error since this
indicates that this pid has been used by other process, so the pid of the
original process is no longer alive.

Differential Revision: https://phabricator.services.mozilla.com/D76692

UltraBlame original commit: 5c89d6071cbbc4bbefa21123e4ff3c6d2d137fe6