Bug 1599858 - Split CompilationInfo::funcData array. r=mgaudet
Split the array of variants into an array of JSFunction and and array of
ScriptStencil. This will let use hold the Stencil generated by the BCE until
the end. In the future we will be able to remove the JSFunction array.

Differential Revision: https://phabricator.services.mozilla.com/D76549

UltraBlame original commit: cfefeb4ca74b11e2433f97add0abf27d0340f4a2