Bug 1598734 - Refine background script toggleProfiler and captureProfile commands; r=canaltinova
This patch handles the issue where a user sends two commands two toggle the profiler and/or
capture. This gets the profiler UI in a weird state that usually is just because the UI
is hanging and lagging. This makes the profiler pause state into a "capturing" state.
I did not include a test with the behavior change, as I was worried about intermittent
failures on asserting this behavior.

Differential Revision: https://phabricator.services.mozilla.com/D75850

UltraBlame original commit: 147bf269db009ffc7dc30f188d70ee4d0d09a1e9