Bug 1641789 part 1 - Change TruncateDoubleToUInt32 input from ValId to NumberId. r=evilpie
Use GuardIsNumber for binary bitwise operations so that the stub works for int32
too. The IC code may be slightly less efficient now for int32 values because they
are converted to double but this is consistent with other code and helps Warp a lot.

Differential Revision: https://phabricator.services.mozilla.com/D77420

UltraBlame original commit: 102cb2082b8a68ed2009be0137c5462c23ffdb56