Bug 1637235 - Implement a ring buffer for audio data. r=padenot,jya
Implement a ring buffer that is not thread-safe and preallocate its internal buffers. The intention is that the internal data is preallocate to any thread and then read/write operations will take place in a single thread using the memory in a ring manner.

Differential Revision: https://phabricator.services.mozilla.com/D74882

UltraBlame original commit: 7600c1bbd621e9ba857d419706b11e8b8ecbdfa0