Bug 1641468 - Add platform message stub test. r=ochameau.
This patch adds a mochitest that will be responsible to
generate and update platform messages stubs, which is what
we already have for all the other type of resources we handle
in the console.
This means the mocha test asserting the rendering of platform
messages can use those stubs instead of hand-written stubs we
have at the moment.

This patch also adds a `getCurrentTestFilePath` function in shared-head.js,
which we then use in the different stub generators.
We also use CHROME_URL_ROOT to require the stub-generator-helpers file.

Differential Revision: https://phabricator.services.mozilla.com/D77216

UltraBlame original commit: 2a670d8c9c3e98406108cafc1f281c6c2d90a583