Bug 1643201 - servo: layout_2020: Add initial support for getComputedStyle.
This implementation is more-or-less on par with the one from layout_2013
and in some cases better. There are still some cases where we don't
return the correct "resolved value," but this is enough to test
animations and transitions.

Depends on D78190

Differential Revision: https://phabricator.services.mozilla.com/D78191

UltraBlame original commit: 96fbb7b812dad0147fbacf8fc17de9d6a1116040