Bug 1628948 - Dispatch mousedown asynchronously on a WM_NCLBUTTONDOWN. r=NeilDeakin
On other platforms we get a mousedown event when the non-client area is clicked,
but not on Windows, where handling the WM_NCLBUTTONDOWN message breaks dragging
the window. This dispatches the mousedown event like other platforms, but on
the next tick.

Differential Revision: https://phabricator.services.mozilla.com/D77419

UltraBlame original commit: cb3bea25854a0893beaa9f03ea3d76f9494ab703