Bug 1641600 - Remove check that bytesSent >= packetsSent in pc.js; r=ng
With RTX and transport-cc enabled, it is possible to send padding packets prior
to any media being sent (see comments in RTPSender::SendPadData). But padding
bytes are counted separate from payload bytes and so do not show up in
bytesSent, which means that is possible that we've sent packets, but not bytes,
as far as the stats are concerned. This leads to frequent intermittent failures
while running tests on Android hardware.

Depends on D77815

Differential Revision: https://phabricator.services.mozilla.com/D78287

UltraBlame original commit: 297d2efe3bb859ac149faa038910131047489b15