Bug 1643706 - Make spatial_tree.print() work on Android too. r=kvark
Instead of using the default TreePrinter implementation which sends output
to stdout, collect it into a Vec<u8> and use the debug! macro to log it,
which sends it to a more useful destination (logcat) on Android.

Depends on D78536

Differential Revision: https://phabricator.services.mozilla.com/D78537

UltraBlame original commit: 4e89164de2d6bbefbbabcce580658a69d12b149b