Bug 1637680 - Make softDisable setter async. r=mixedpuppy
Some tests rely on the order of execution of the blocklist code which cannot be
guaranteed if we don't wait for the enabled state to settle.

In particular, when calling `_onUpdate` in
`AddonTestUtils.loadBlocklistRawData` in [0] we implicitely assume that all the
actions deriving from updating the blocklist settle before the promise is
completed, this is not true for the enabled state.

Currently the test succedes because `ExtensionPermissions.get` is not actually
async in most cases, this is about to change in the next patch.

[0]: https://searchfox.org/mozilla-central/rev/559b25eb41c1cbffcb90a34e008b8288312fcd25/toolkit/mozapps/extensions/internal/AddonTestUtils.jsm#925

Differential Revision: https://phabricator.services.mozilla.com/D77937

UltraBlame original commit: 116a2884768441d254b91a55206052ea6b948844