Bug 1643685 - Use name instead of test_name when finding the metric test name. r=mozperftest-reviewers,tarek
This patch fixes the test name being used for the App-link test in CI (regression). The patch also adds a new feature to ScriptInfo which checks to make sure that the test doesn't have any extra unknown keys in it. If it does, a failure is raised. This change is to attempt to help prevent these kinds of regressions from happening again. A new test is added for this functinality as well.

Differential Revision: https://phabricator.services.mozilla.com/D78516

UltraBlame original commit: bac92eaf4731bd8fbb1a1f07c270ac6c4ee34dc6