Bug 1637845 - Add a new function 'copy_contents' to mozfile r=glob
When you have a directory src and you want to move all the contents
to dst, this is surprisingly annoying. shutil.move doesn't operate
on globs. shutil.copytree does exactly what we need *except* it
requires the destination to not exist. In Python 3.7 it gained the
ability to not care about the destination existing but... we need
to handle the pre-Python 3.7 case.

Differential Revision: https://phabricator.services.mozilla.com/D76083

Depends on D75698

UltraBlame original commit: b459114e8008bd289607361b39c2e25b1fa9cb3f