Bug 1630706 - Part 1: Allow non-shared ArrayBuffers for Atomics operations. r=jorendorff
The TypedArray data extraction had to be moved from `AtomicAccess` into the
actual operation to account for the case when a TypedArray using inline data
is moved by the GC due to side-effects triggered by type conversions within
the operation. Part 4 contains tests to cover this case.

Because the TypedArray data is no longer passed to the operation, we can't use
it anymore to determine the data type from the `SharedMem<T>` parameter. Instead
pass an `ArrayOps<T>` instance to the operation, so we can determine `T` within
the lambda function. (Templated lambda functions are only available starting
with C++20!) The compiler should be able to optimise away the stack allocation
for this instance, so this won't incur any performance costs.

Differential Revision: https://phabricator.services.mozilla.com/D76319

UltraBlame original commit: dd1ec565b69a673e08f1d8a5855d12a2da64ba16