Bug 1599160 - Temporarily don't wait for the zoom messages in a DAMP test. r=nchevobbe a=NPOTB. CLOSED TREE
DAMP refreshes the page that it loads, but doesn't clear the stylesheet
cache, so the second+ time around some of the zoom messages will not be
sent.

This is actually a wider issue, that I plan to fix in bug 1645180.
Meanwhile don't wait for these messages.

Differential Revision: https://phabricator.services.mozilla.com/D79352

UltraBlame original commit: edb0e17f3c984994dec1ba84f509e78428de022b