Bug 1640034 - Ensure that we don't bail when we need the parent frame created. r=jorendorff
The inline comment explains most of this, but essentially because we bail out
of the FrameIter loop early when we find an asyncStack, we can end up with
cases where the a Debugger eval's parent frame either hasn't been populated in
the LiveSavedFrameCache at all, or else _has_ been, but is not instantiated
in the correct realm.

This patch avoids that by no longer bailing when asyncStacks are found, and
instead traversing the whole stack still to make sure that the invariants
for the LiveSavedFrameCache are preserved properly.

Differential Revision: https://phabricator.services.mozilla.com/D76426

UltraBlame original commit: eb657b731e1a58a48e9d7be20fefd9fdaf3e2d5d