Bug 1644966 - Change parameter of ContentProcessListener to a function. r=jdescottes.
It used to take an object with only 1 property, a function. The way it was
called in the webconsole actor made it harder to follow than it actually need.
We take that as an opportunity to convert the function to a class.

Differential Revision: https://phabricator.services.mozilla.com/D79294

UltraBlame original commit: 0da6f494ad2c5515c91c5d4cea562daf69b48890