Bug 1644985 - Clean up WeakRef and FinalizationRegistry data when cross compartment wrappers are nuked r=sfink
This crash happens because we try and clean up the map from target to WeakRef in the WeakRef finalizer, and the target can be dead by this point if it's a nuked CCW (before it is nuked the CCW ensures this sweep order does not happen).

The fix is to fix up the map when CCWs to WeakRefs are nuked. Fortunately there's already a hook where the GC is told about this.

The same issue applies to FinalizationRecordObjects. This fix is slightly different because they don't have a target pointer so we can't find the map entry. Instead we clear the record and cleanup happens later when it gets swept.

Differential Revision: https://phabricator.services.mozilla.com/D79533

UltraBlame original commit: 75b4198a731db7dcbfc6280bf25639b083e36cca