Bug 1646206 - Fix a vsnprintf lint in ClearKeyUtils.cpp r=bryce
We have a clang-plugin check that discourages raw `vsnprintf`, but for some reason it doesn't catch this before clang 11. I _think_ it's related to earlier versions not being able to see that `len` is a constant, but I'm not really sure.

Differential Revision: https://phabricator.services.mozilla.com/D79918

UltraBlame original commit: 2abc34452a31cc985b2bce50523b83cb6acd3457