Bug 1635273 - Make FT_LCD_FILTER_H a relative rather than absolute path. r=froydnj
This helps get sccache hits across source directories.

The previous presentation guaranteed the version of `ftlcdfil.h` in
the tree, while this unorthodox presentation only ensures that the
`freetype2` directory has parent `modules`.  This is at least unlikely
to occur in random `/usr/include` directories.

If this turns out to be an issue it would be possible to copy
`ftlcdfil.h` into the object directory, potentially with a unique
name, and then reference that (with a relative path).

Differential Revision: https://phabricator.services.mozilla.com/D73759

UltraBlame original commit: ec18fc1e8fe09ccb4e9848cc832d1e4277fe3eae