Bug 1642591 - Fix image/test/unit/test_private_channel.js. r=tnikkel
This test fails already on central if you change this line:

  https://searchfox.org/mozilla-central/rev/027893497316897b8f292bde48dbb6da2391a331/image/test/unit/test_private_channel.js#93

by:

  Ci.nsIReferrerInfo.EMPTY

Because the previous part of the test populates the image cache. The
test wants to check that the channel for the image load is properly
flagged as private and thus that the http cache is partitioned
appropriately, thus clearing the image caches seems sane.

While at it, also fix it so that we send a valid image instead of base64
text, though that change is not technically needed so feel free to ask
me to remove it.

Differential Revision: https://phabricator.services.mozilla.com/D79848

UltraBlame original commit: 82d4df065e00f0c9e75388e11679108c3ee26cbb