Bug 1643607 - Fix the timeout issue of the test browser_aboutCertError.js in Fission. r=johannh
The timeout happens when we get an old element of the 'errorCode'
element in the test code. In this patch, we fix this issue by getting
the element every time we check the condition of the 'errorCode'
condition.

Differential Revision: https://phabricator.services.mozilla.com/D80214

UltraBlame original commit: ce45f7273cc0bf01e4432f6791d1588de5f17444