Bug 1637475 - Flush layout when creating background pages so the browser element gets its frames created. r=mixedpuppy
Before bug 1584935, the frame for the browser element was created
directly after the .appendChild call, but after that we need to wait
for layout to happen in order to guarantee it.

The code was duplicated in ExtensionXPCShellUtils.jsm so I also put the
wait there.

Differential Revision: https://phabricator.services.mozilla.com/D76601

UltraBlame original commit: 0542461e3da7318b777e2e606dbbfff69c298924