Bug 1647065 - Force non-legacy ExtensionPermissions tests to use rkv. r=robwu
If the json bakend is enabled (e.g. in beta as of the time of writing)
we need to force the tests that use ExtensionPermission to run in rkv mode. The
non-rkv mode code path will be tested in xpcshell-legacy-ep.ini.

Differential Revision: https://phabricator.services.mozilla.com/D80378

UltraBlame original commit: 8a3d6d3ca0b79dbea39c573fb1bac8848ed63bfc