Bug 1625044 - Do not run the profiler private browsing test on linux; r=julienw
I spent some time looking into this, and the Linux runs when opening up the
popup and the private browser are not working consistently. The popupshown
and popuphidden events are not firing consistently. It's probably good enough
to just disable this test on Linux, as it's mainly a UI test, and it works
on the other platforms.

Differential Revision: https://phabricator.services.mozilla.com/D81602

UltraBlame original commit: 5a8c95e9db1c33ce2072ea4366efaa0759f009a8