Bug 1577639 - Do not attempt to clean up frames for non-OOM conditions. r=arai
The core issue is that we run removeFromFrameMapsAndClearBreakpointsIn(cx, from)
even for non-OOM paths in this function, which is undesirable because that will
remove the breakpoints for evaled scripts. In the non-OOM case there are no
Debugger.Frame objects left in the frames map for the "frame" list, so there
is no reason to actually call it in that case.

This patch merges the two OOM handlers together for simplicity, though it
is not technically necessary, it seems like the very minor increase in code
run if there is an OOM is worth it to keep things centralized.

Differential Revision: https://phabricator.services.mozilla.com/D80762

UltraBlame original commit: 3310725a4fcd81e98eb5577c307c5c9915aed6de