Bug 1647342 - Part 2: Rely on onResumeFrame to handle marking all frames as on-stack. r=arai
It is guaranteed that this will already be populated because in 1539654 we
started properly preserving the debuggee-ness of frames across suspend and
resume. With slowPathOnResumeFrame guaranteed to run, we no longer need to
try to hande handle the case where the generator has been resumed, but
the Debugger.Frame's inner data was not populated.

Differential Revision: https://phabricator.services.mozilla.com/D80735

UltraBlame original commit: ac9788d58fe347bf6dfc5b6ec962fffdac77e42f