Bug 1648703 - Use a custom vector growth pattern in PrimitiveList. r=gw
This greatly reduces the number of vector reallocations happening while building primitive lists. On the difficult cases like youtube front page the reduction is a bit more than 50%, and more in other pages I tested. More importantly it dramatically reduces the amount of the most expensive of these reallocations which are when the vector is starting to get large.

Differential Revision: https://phabricator.services.mozilla.com/D81725

UltraBlame original commit: 8deaca01428d9ff281089c80d15c1b7fbfb25670