Bug 1645710 - Store Scope's data pointer in the cell header to avoid race during compacting r=jandem
Currently Scope stores the pointer to the enclosing scope in its cell header.  This can be both read and written to while updating scopes during compacting GC.  The result is the same regardless of the order the operations, since the read is to check the forwarded flag (which is always clear in this case) and the write doesn't change the value of the flag.  I don't think this is causing problems at the moment, but it is undefined behaviour and should be fixed.  This race was previously suppressed.

The patch stores the data pointer in the cell header which doesn't need updating during compacting GC.

Differential Revision: https://phabricator.services.mozilla.com/D81682

UltraBlame original commit: b0cc716aa5c034ee11a92c5bcd93655b11efd407