Bug 1649710 - Fix EnvironmentFront constructor. r=jdescottes.
The constructor signature was wrong, as protocol.js calls it with a targetFront
instead of the expected "form".
The self-management isn't making much sense, and there's no need for this front
to be different than others.
test_framebindings-07.js, which was directly instantiating EnvironmentFronts is
modified to not do this anymore.
This allow us remove the getBindings method from the front, as it was only used
from this test.

Differential Revision: https://phabricator.services.mozilla.com/D81838

UltraBlame original commit: 7220ff5029d0d11c2e3510351816d15e784ebcb9