Bug 1619953 - P3 - Asynchronous BlobURLInputStream added r=baku
Reading from this stream type should be the preferred way to consume data from blob url. When created in the parent process, it requests blob data locally from BlobURLProtocolHandler. When created in a content process, it makes BlobURLDataRequest IPC call to the parent process. Should be wrapped in a nsBufferedInputStream, because ReadSegments() is currently not implemented for this stream type.

Differential Revision: https://phabricator.services.mozilla.com/D75292

UltraBlame original commit: 4668c99560de26a931c3cc96f2fee432b2898c59