Bug 1648354 - Handle the fact that ToPtr() may return null in the child process. r=jwatt
Normally, if IsNull() is false, we'd expect ToPtr() to return a valid pointer,
but for code that may run on a stylo thread in the child process this is not
necessarily true: if resolving the pointer requires accessing a new shared-mem
block, we can't make that IPC call from the stylo thread. So in this case, we
let ToPtr() return null, and the caller needs to handle this safely.

Differential Revision: https://phabricator.services.mozilla.com/D81242

UltraBlame original commit: 3e78b6a57f062559963f618bb12275dc1431f36b