Bug 1649970 - Correct the index for fetching next available image r=alwu
Suppose the current artwork is <img A, img B, img C> and Firefox is
currently fetching img B, whose index is 1 in the artwork. If the
artwork is updated to <img D, img B> before img B is fetched, then
Firefox will skip fetching a new image since img B is also an acceptable
choice in the new artwork. However, if Firefox fails to fetch img B, in
the current implementation, a new image at (previous index + 1) would be
fetched, as a fallback plan. In the above case, the index is 2, which is
out of the bound of the new artwork.

Instead of caching the previous index of an image in the artwork and
using the the previous index plus 1 as the index of the next available
image, the index of the next available image should be reset every time
when artwork is updated

Differential Revision: https://phabricator.services.mozilla.com/D81960

UltraBlame original commit: d1934bb1c06def3e92d50fc42d76ed0b99803463