Bug 1646378: Factor out ICScript from JitScript r=jandem
When we clone a JitScript for inlining, there is some data that we want to share and some data that needs to be separate. This patch moves the separate data (warmup count, ICEntry array) into ICScript and embeds an ICScript into JitScript. The layout of a JitScript in memory is mostly unchanged by this patch.

Differential Revision: https://phabricator.services.mozilla.com/D80196

UltraBlame original commit: cbe0ca1d37b796bfc41b25bdf7e5ae44ca9ef4c2