Bug 1650239 - Build llvm-dsymutil out of the clang-10 source tree r=firefox-build-system-reviewers,rstewart
The current version hits OOM errors when dsymutil-ing files created by clang 10 and 11.

The choice of clang-10 here is somewhat arbitrary in that it's a fetch job that we conveniently already had in the tree. It doesn't have to be exactly version 10 specifically.

Differential Revision: https://phabricator.services.mozilla.com/D82453

UltraBlame original commit: 75384d97c9df8eadba54e30f6db70659b757c554