Bug 1641256: Don't deactivate remote canvas due to read error when the writer has already failed. r=jrmuizel
This also removes the gfxDevCrashes when an invalid enum is read, because we
suspect these are mainly due to the writer shutting down mid event.
We could only crash when the writer hasn't failed, but because these reads are
in templated code it would mean updating the different event streams with a new
function. If we are still getting high numbers in the deactivation telemetry we
will need to do this to try and track down the problem.

Differential Revision: https://phabricator.services.mozilla.com/D82493

UltraBlame original commit: 50bf5757f9113224d109fad35395add16275c07a