Bug 1626058: Don't add prefix when building a docker-context at graph generation time; r=glandium
We were building a context with a top-level directory matching the image name
during graph generation. This is not accepted by docker build. This didn't
cause issues since we only used it for digest generation. Since we are
switching to use the decision generated context, change this to match the
context being generated when building an image.

Differential Revision: https://phabricator.services.mozilla.com/D77841

UltraBlame original commit: d5022eddaa923c482622c8d731ca13129099e0a5