Bug 1650837 - Part 3: Destroy active actors when they are unregistered, r=kmag
This is needed specifically for tests, which need to be able to unregister &
re-register the same actor with different configuration options. With
JSWindowActors, we can create new windows after each actor is unregistered, but
we can't do this with in-process JSProcessActors.

Differential Revision: https://phabricator.services.mozilla.com/D82450

UltraBlame original commit: a14f2e220852af610b824cb667901160296686cb