Bug 1650224 - Drop support of testPath from backfill action r=tomprince
This optional mode to specify a test path when backfilling has been broken for a while
and is not being used to make it worth fixing. It is a mode
originally designed to be used with the test-verify tasks.

Differential Revision: https://phabricator.services.mozilla.com/D82722

UltraBlame original commit: c96d73d17b9a5dc022df57bc4f41d265eafcc099