Bug 1651641 - `servo/components/style/build.rs` reads `PYTHON3` configuration value in addition to environment variable r=glandium
The previous logic would check the value of the `PYTHON3` environment variable, which is not set by the Firefox build system, so this build script tends to just use the `python3` at the tip of the `PATH`, regardless of which value you've configured for it. Instead, it should prefer to read from the `OBJDIR` if one is present.

Differential Revision: https://phabricator.services.mozilla.com/D82981

UltraBlame original commit: cbfe3335cfe59e5222a57bb81b4aa02e9eed5efa