Bug 1648839 - Fix two issues with notifications for media feature value changes. r=nordzilla
We need to notify unconditionally because even if we didn't have
stylesheets, we could have responsive content which needs to change
source.

We need to notify the document even if the pres shell is not
initialized, as it might be our last chance to notify the responsive
content. This happens for printing, and makes my srcset test fail.

MANUAL PUSH: Somehow phabricator still thinks it needs review, even
though it was accepted.

Differential Revision: https://phabricator.services.mozilla.com/D81778

UltraBlame original commit: cff48285254bc9a82fa3f708ae4143639953f10e