Bug 1649980 - part 6: Redesign `WSRunObject::MaybeReplacePreviousNBSPWithASCIIWhiteSpace()` r=m_kato
It's called by `WSRunObject::InsertText()` and `WSRunObject::InsertBreak()` and
it has 2 parts.  One is, considering whether previous char is an NBSP or not
and whether it should be replaced with an ASCII white-space if it's followed
by visible character.  The other is, doing the replacement.  The latter code
is enough simple.  Therefore, we can copy them into the callers.  Then,
we can move the check logic into `TextFragmentData`.

Depends on D82698

Differential Revision: https://phabricator.services.mozilla.com/D82699

UltraBlame original commit: 6fedbc9d3130fce2b95154578b9a8bb9fe26e7e0