Bug 1652205: Update one-liner comments for printed page frame classes. r=jwatt
DONTBUILD because this is a comment-only change.

Before this patch, these comments all mentioned "simple page sequence frame",
which is no longer a thing (it's been renamed to nsPageSequenceFrame). This
patch fixes that and corrects/clarifies these comments while we're at it.

(Note that these comments/classes may change a bit, as part of future patches
for the "pages-per-sheet" feature.  This patch here is partly to get them to a
coherent starting state, for that work to build from.)

Differential Revision: https://phabricator.services.mozilla.com/D83233

UltraBlame original commit: beee8123ffb37ef647610e4b3df0533b9d104bb9