Bug 1651645 part 6 - Fold ensureDenseInitializedLengthNoPackedCheck into ensureDenseInitializedLength and call from ensureDenseElements. r=evilpie
The old code in ensureDenseElements marked non-packed before extending the
elements capacity. That's a problem because it can be the statically-allocated
EmptyObjectElements and we can't set the flag on that.

It's clearer to do the non-packed check in ensureDenseInitializedLength.

Depends on D82962

Differential Revision: https://phabricator.services.mozilla.com/D82963

UltraBlame original commit: 726d3e7fd0428cf47be8bd3cb7931551f8db3c77