Bug 1651605 - Disable Service Worker support in DevTools Debugger r=Honza
We ran an experiment to assess the impact of the feature.
We could verify a significant performance impact on the opening times for the debugger.
It also led to a decrease in overall devtools usage and retention.
Therefore we are disabling the feature until we can implement it with better performance.
This should be implemented with https://bugzilla.mozilla.org/show_bug.cgi?id=1651522

Differential Revision: https://phabricator.services.mozilla.com/D83312

UltraBlame original commit: 4c48ff27577f197ba45b4476f831adc55804046c