Bug 1645123 - avoid trapping instructions on Skia debug asserts. r=jrmuizel
clang may potentially insert trapping instructions when __builtin_unreachable() is used.
We need to ensure we do not use this intrinsic when asserts are disabled. To work around
this, we add a condition that is checked before entering the body of SK_ABORT.

Differential Revision: https://phabricator.services.mozilla.com/D83841

UltraBlame original commit: ab2f42930fb0a048336b59d0fe6568ca0430e02e