Bug 1648453 - Don't try and store the incumbent global directly r=sfink
The previous patches store the incumbent global directly, which is going to be a problem if it's in another compartment. It needs to be wrapped into the current compartment. But apparently that can be problematic since you don't know how far to unwrap it to get the original gobal object back (it might itself be a wrapper). The solution used for promises is to store a CCW to another object with the same global. This patch applies the same fix to FinalizationRegistry.

Differential Revision: https://phabricator.services.mozilla.com/D83831

UltraBlame original commit: fcb43544e8a2539f33f83ac771005304ee4425ad