Bug 1612831 - [marionette] Switch to top-level frame before checking current URL. r=marionette-reviewers,maja_zf
The navigation command is always executed on the top-level
browsing context. As such the check for a possible load
event should not be done for the currently selected frame.

At the same time the switch to frame call can be moved
to the parent process.

Differential Revision: https://phabricator.services.mozilla.com/D80619

UltraBlame original commit: b066b09fe903bf51bf85c0726839905a99dca45e