Bug 1652870 - permit DUMP_SYMS to be used when doing Android builds; r=mhentges
It's not immediately obvious to me why we limit this check based on target;
I guess once upon a time there was no ELF or Mach-O support, so we wanted to
ensure people didn't shoot themselves in the foot. If that's the case,
testing indicates that Mach-O support isn't quite ready for prime-time and
we haven't got all the bits straightened out for our normal Linux builds.
So we're just going to enable it for Android here.

Differential Revision: https://phabricator.services.mozilla.com/D83560

UltraBlame original commit: e4cd31ea3ca1b6535f84e5e22794dabd056499a0