Bug 1653988 - Extend the loading of dump fallback for Remote Settings to cover more indexedDB failure areas. r=leplatrem
This extends the current try/catch to cover writing of the dump to the database and reading it back again. It does not cover the verify signature section, as it is expected that it will need to be known if the signature fails.

Differential Revision: https://phabricator.services.mozilla.com/D84234

UltraBlame original commit: 504fea031a58e6a930329deb1979e168d1a8e957