Bug 1654569 - Keep BrowsingContextGroup alive throughout the process switch, r=farre,annyG
This builds on the new API added in bug 1652085 to reduce the chance of a
different BrowsingContextGroup instance being used for process selection at the
start of a process switch, and BrowsingContext creation at the end.

While this would probably not be a serious issue right now, as we always have
only a single "extension" process, it could become an issue in the future if the
specific group specifier is used in more places.

Differential Revision: https://phabricator.services.mozilla.com/D84549

UltraBlame original commit: 4388af0c7d953456560be39fe8289da18a94a53e