Bug 1654439 - Part 2: Factor out some urlbar-specific things from SearchOneOffs into UrlbarSearchOneOffs. r=harry
The base class retains the `popup` property, and now the `view` property is
entirely in the subclass. I defined some new methods so that the subclass can
easily customize behavior for a view that's not a popup. I used the name "view"
in these method names to refer to a generic view/panel. I didn't rename `popup`
`view` because the base class still assumes that `popup` is a `menupopup`.

We could/should probably get rid of the "compact" concept and instead specialize
the behavior even further in the subclass, but for simplicity's sake I left that
for now.

Depends on D84784

Differential Revision: https://phabricator.services.mozilla.com/D84785

UltraBlame original commit: c03e14b7c43066793396e88f184f1280a10f8603