Bug 1377100 - Store sync metadata in the LoginManager. r=lina,MattN
This is done so that if the login manager loses all data sync is "reset",
meaning will do a full reconcile with the server and pull all sync records
down and apply them locally - ie, sync will be ble to recover passwords for
many users.

The sync GUID is stored encrypted, so that even if the login manager data
isn't lost, but the encryption key is, it will still reset and recover.

This will also make restoring an older logins.json a little more reliable for
sync - if an old version is restored then sync should "rewind" itself to the
state it was in the backup.

Differential Revision: https://phabricator.services.mozilla.com/D82663

UltraBlame original commit: 0d918902af47e02efbe37253d32a0e0292338c2e