Bug 1655474 - Don't crash when trying to resume print while the page is closing. r=jwatt
This is a pre-existing issue that could already happen before the
regressing bug but seems worth addressing anyways.

If we're destroying we're definitely not going to be able to print, so
return an error rather than crashing in funny ways otherwise.

Depends on D85038

Differential Revision: https://phabricator.services.mozilla.com/D85039

UltraBlame original commit: 0506dfe2c979491c658de924b0ddf26e796cd099