Bug 1653263 - Introduce a page-icon relation expiration and use it to expire icons when storing new icons for a page. r=Standard8
Until now, when icons for a new page were stored, we removed the old page-icon
relations based on the icon expiration. But a single icon can be in use from
many pages and it's expiration gets bumped for any of them. This means that if
an icon has even just a single valid page in use, all of its relations may
survive forever.
This patch introduces an expiration on the page-icon relation instead, so that
the relations themselves can be expired when new icons for a page are loaded.
Icons without relations (provided they are not root icons) will then be expired
by Expiration, Maintenance or History.

Differential Revision: https://phabricator.services.mozilla.com/D83952

UltraBlame original commit: 623ac5b0c08511b899d2695b5c4e9990a254c512