Bug 1654295: Use Result as a return type where possible in IOUtils.cpp r=barret,Gijs
This changes most private methods of `IOUtils` to use the `Result` monad as a
return type instead of a bare `nsresult`. This improves consistency of all of
the public method implementations, and will allow for the shared patterns to be
abstracted in a follow-up patch.

Differential Revision: https://phabricator.services.mozilla.com/D84734

UltraBlame original commit: ef65a442f85265d4240098577a9a50bc42492d22