Bug 1654295: Reduce boilerplate in IOUtils failure paths r=barret
This change completely overhauls the way errors are handled in `IOUtils`. With
the introduction of the new `IOError` type, a useful error message is paired
with an `nsresult` at the error site. This provides erroneous callers with more
information about what went wrong, while improving consistency with mapping
errors to `DOMExceptions`.

For error sites where it is not immediately clear what went wrong, messages can
be omitted. A default error message will be filled in corresponding with the
wrapped `nsresult` when the operation is rejected to the calling JavaScript.

Differential Revision: https://phabricator.services.mozilla.com/D84736

UltraBlame original commit: 63c355562fa6650c8c84b5c169300458951af074