Bug 1653248 - Remove AbstractScopePtr from ScopeCreationData r=arai
Replace the enclosing scope field with a Maybe<ScopeIndex> field. When the
enclosing scope is an existing concrete scope, the CompilationInfo data will
be used instead.

One source of complexity is that scripts in the self-hosting global use the
existing empty global scope and indicate this with a placeholder type in the
gc-things array. In this case there is no ScopeCreationData and we return
mozilla::Nothing for the scopeIndex. When ScopeCreationData::enclosing
computes the actual scope, it handles this case.

Differential Revision: https://phabricator.services.mozilla.com/D84889

UltraBlame original commit: 1e4afa0fa38150f2425aeea78cb44d0fd32ef3ec