Bug 1652278 part 1: Add stub implementation for new frame class "PrintedSheetFrame", to represent an individual printed sheet of paper. r=TYLin
This patch is just to get the "new file/frame class" boilerplate out of the
way.  As of this patch, this frame class *is* compiled, but it doesn't do
anything and it's never instantiated. The next patch in this series will make
us actually start using it at runtime.

Differential Revision: https://phabricator.services.mozilla.com/D83456

UltraBlame original commit: bcd829a3b3f9e2c837d3bb1806083dbc9cd61e01