Bug 1656348 - Add source support to FormHistory. r=adw
Add an optional source property for FormHistory that can be defined when bumping,
removing of querying values.
This can be used by consumers to partition FormHistory based on specific sources,
for example the name of the engine used to execute the search.

Differential Revision: https://phabricator.services.mozilla.com/D85489

UltraBlame original commit: 6fc161cca448cf08de943898fe137558d78fd974