Bug 1656141 - Remove dav1d.rc. r=achronop
RCFILE is only actually used when the moz.build that contains it defines
a binary, which is not the case for dav1d, which ends up in gkmedias.dll.
Which also means that moving the definition to gkmedias would also not
make sense, since all dav1d.rc does is add descriptors to the dll that it
contains dav1d and what version, but gkmedias.dll contains other things
too.

Differential Revision: https://phabricator.services.mozilla.com/D86152

UltraBlame original commit: de8899453150475aa6337f352e8f7122f1482a72