Bug 1652445 - Remove skipped & outdated test browser_inspector-mutations-frameload.js r=ochameau
Depends on D85598

This test relies on mutations which are no longer emitted for most navigations and which will no longer be emitted at all with the next patches.
It has been disabled for 9 months and should be removed.

Differential Revision: https://phabricator.services.mozilla.com/D85599

UltraBlame original commit: a1d88e10564449cf55525d959a60edfeb85163d1