Bug 1658330 - Start a new query to show top sites when the input is empty and search mode is exited by backspacing. r=harry
We just need to start a new query on backspace, even when the input is empty.

I noticed that unlike backspace, clicking the close button will always start a
new query except when the view is closed. That means the close button was
already doing the right thing. It also means that the close button will not open
the view and show top sites when the view is already closed; however, with this
patch, backspace *will* open the view and show top sites if the view is closed.
That seems right to me despite the inconsistency since inputs in the urlbar
(including backspace) always cause the view to open.

Other changes:

* Make the `click_close` test task more similar to the `backspace` task: Open
  the view both with and without a string instead of only with a string
* Some calls in the test didn't have an `await` like they should (maybe the
  source of reported intermittent failures?)

Differential Revision: https://phabricator.services.mozilla.com/D86686

UltraBlame original commit: 427683ea28aca31d17fd8ca7774e3b8bd6fc05ab