Bug 1657449 - Add "Show tab" button to about:webrtc; r=ng
This adds the current browserId to the internal stats report. The peer
connections are sorted by browserId, and a "Show tab" button is added that will
select the tab associated with the peer connection to make it easier to keep
track of which peerconnection is associated with a tab.

Differential Revision: https://phabricator.services.mozilla.com/D86699

UltraBlame original commit: de259037a7fff6d8d6cbbae9cd4082847357df8d