Bug 1657717: pipenv should use same version of python as currently active, if possible r=firefox-build-system-reviewers,rstewart
Pipenv enthusiastically checks pyenv for possible Python installations, and it seems to always want to use the most
up-to-date version possible. This can lead to issues if the version of python used for the regular venvs is older
than another version of python that exists on the machine (such as if the system python is 3.8.2, but pyenv has
3.8.3 installed).

This patch addresses this by scoping pipenv to match the currently-running version of python (if major versions line
up).

Differential Revision: https://phabricator.services.mozilla.com/D86448

UltraBlame original commit: d141ba87d0c1d7420bf2bc0fa686b84aa51437e4