Bug 1658225. Ensure that WidgetWheelEvents produced from PinchGestureInputs always have a non-zero deltaY. r=kats
See bug 1648489 for why we need this.

If we don't send a pinch we don't want to update the last scale so that when we do send an event it incorporates the full scale change from the last time we sent an event.

Even if aScale != mLastScale in SendPinch the computed deltaY ( = 100 * (aScale-mLastScale) * widgetscale) can be 0 seemingly because of floating point weirdness I guess.

Depends on D86495

Differential Revision: https://phabricator.services.mozilla.com/D86496

UltraBlame original commit: bf72f2e5024ba404c624e3aedbaa11d445a81202