Bug 1656261 - Disable StartupCaching of Omnijar zip central r=froydnj
This is a speculative fix for a crash we're seeing due to xul.css ostensibly
not existing. The theory is that xul.css does in fact exist and the cached
zip central for the omnijar is simply corrupt in some way. If it is corrupt in
this way, then there is a bigger issue, and we need to investigate deeper.

However, the benefit of this approach is that it is a very small and contained
patch which should be simple to uplift.

Differential Revision: https://phabricator.services.mozilla.com/D86829

UltraBlame original commit: 6eeb525ee65a9193e1098fcb2c22ac1860374461