Bug 1633031 Part 1 - Stop considering available block-size when wrapping flex lines in multi-line column-oriented flex container. r=dholbert
Although this goes against the suggestion of the sample algorithm in the
spec, this brings the layout of the multi-line column-oriented flex
container in paginated context closer to the layout with no
fragmentation. This also matches Google Chrome's behavior.

Tweak flexbox-unbreakable-child-2.html so that it matches the rendering
as of this patch. Change the height of `.muticol` and the removal of
border in `flexContainer` to let two flex items fit exactly into the
flex container's height, making reference easier to write.

Differential Revision: https://phabricator.services.mozilla.com/D87107

UltraBlame original commit: b1220dc3ec0982444cdee534a265926a977daf24