Bug 1659151 - Mark wrapper rooters in mark slice if we might start sweeping in this slice r=sfink
The idea behind removing marking of wrapper rooters from the mark phase is OK
except we rely on marking being done by the time we start sweeping (we can't
yield between the end of marking and the start of sweeping) and if we delay
marking these wrappers until the sweep phase we may end up with marking work to
do at this point.

The fix is to mark these in mark slices that may end up sweeping (mostly this
doesn't happen because we yield before we start sweeping). Then the marking
gets done by markUntilBudgetExhausted before we get to the sweep phase.

Differential Revision: https://phabricator.services.mozilla.com/D87459

UltraBlame original commit: 098f60d27d39013e8f690c9fc0ccb025c0aad6ad