Bug 1622995: Expose video controls to a11y even when visually hidden. r=MarcoZ,Gijs,heycam
This required a change to vtt.jsm because it previously used clientHeight to determine whether the controls were showing.
Now that the controls are visually hidden but not display: none, that doesn't work because clientHeight can be > 0 even when the controls are visually hidden.
The hidden attribute is now used instead.

Differential Revision: https://phabricator.services.mozilla.com/D87403

UltraBlame original commit: 173c59907b33d4f35b1699521d5f02f8c47cd75b