Bug 1656229 - replace a computation by a constant load.  r=jseward
This is technical debt from the initial SIMD work - it's faster to load a
constant here than to compute it.  Benchmark data not forthcoming, but
elsewhere i've found that we break even at two instructions and it's
better to load the value than to compute it in three.

Differential Revision: https://phabricator.services.mozilla.com/D87306

UltraBlame original commit: 3da7d219a9e4e4f6e23892e2ef27192de1954f9d