Bug 1630809 - read attribution configuration from github r=aki
The partner attribution config is stored in the same repository as the repo manifest for partner repacks, but all in attribution_config.yml instead of default.xml. This extends the existing support for using the Github API to read files to retrieve and process the attribution config.

Differential Revision: https://phabricator.services.mozilla.com/D87728

UltraBlame original commit: 6eae524a8eec354c282b3341aa7ae83d6a37e972