Bug 1661783 - Disable `test_rst.py` on Linux r=ahal
This test begins to fail due to a `pip` installation issue after bug 1660351. It's already disabled on macOS (also due to what the comment says is a `pip` installation issue), so it's very possible the custom `requirements.txt` that this test installs is deficient in some way that surfaces in certain `virtualenv`s under certain circumstances. I can't diagnose the failure, but what I have seen is that bug 1659539 will fix that problem entirely with no extra intervention required, so we should be able to re-enable this for Linux (and maybe macOS as well?) when that patch lands.

Differential Revision: https://phabricator.services.mozilla.com/D88653

UltraBlame original commit: a47ed8d910807ce7ee066707d510e9b8cd349cb6