Bug 1243463 - make `sEmptyTArrayHeader` const; r=mccr8
We fixed what I think is the lone instance of writing into the (empty)
header in `SetLength` because it was causing TSan violations, so we should
be clear to make this `const`.  This change is not terribly effective on its
own (cf. the `const_cast` required to make this work at all), but in the
next patch, we can rig up `sEmptyTArrayHeader` to be surrounded with "guard
pages" and make rogue accesses off the array header a little more protected.

Differential Revision: https://phabricator.services.mozilla.com/D88657

UltraBlame original commit: ac566cd8ff453656c996a3b792579e370a8b7a68