Bug 1662426 - When destroying an HTML <canvas> frame, reset its print callback data. r=jwatt
This ensures that we call the callback again if the canvas becomes
displayed later. This is important because a callback may run after
destroying the frame tree, and in that case some canvas operations may
fail. Without this there's no guarantee that the callback runs ever
again.

Depends on D89111

Differential Revision: https://phabricator.services.mozilla.com/D89112

UltraBlame original commit: dfddd9655decff96386968d3ba52b1faa1629426