Bug 1661030 - Refactor SubDialog and TabDialogBox to support passing multiple window arguments. r=Gijs,preferences-reviewers
This also updates the behavior of the allowDuplicates check:
Instead of only checking if the top dialog is a duplicate, it will now check the whole stack and
skip the dialog open if a duplicate URL is found.
This fixes an issue where callers could alternate between dialogs to bypass the check.

Differential Revision: https://phabricator.services.mozilla.com/D88422

UltraBlame original commit: 9a10517ec1001aa642d7e7fc4135ed9fb53661ee