Bug 1659423 - Also move wayland popup if getting sane popup dimensions. r=stransky
What is going on is that nsComboboxControlFrame first positions the 0x0
widget, then sizes it. But when moving a 0x0 wayland popup, we just bail
out in NativeMoveResizeWaylandPopup().

So unless we also change the offset in some other way, and hit this
code:

  https://searchfox.org/mozilla-central/rev/dc30fc92f2d02abe716426df7eff0f0bbb907da2/view/nsView.cpp#335

when updating the view tree (which doesn't happen with the maximized
window, and actually it's a bit questionable that it happens without it,
there may be some pre-existing coordinate space confusion there) we'll
end up in a plain Resize() rather than MoveResize(), and never position
the widget at the right place.

This doesn't happen for regular web content because the new print dialog
lives in the parent process (which uses nsComboboxControlFrame) rather
than on the child (which uses nsMenuPopupFrame).

Differential Revision: https://phabricator.services.mozilla.com/D89320

UltraBlame original commit: de2ff8f555f0bf64418b6693fc3fd75fa6379b90