Bug 1663411 - Improve the speed of find-in-page when ending a pseudo-element mid-match. r=jfkthame
So what's going on here is that there are pseudo-elements that contain
stuff like content: "Note" or content: "Example".

So when you're searching for "entry", we find the `e`, and then all the
text nodes afterwards are of course in a different subtree... So we end
up ignoring them, but after having iterated through all the DOM first.

To avoid this, change a bit the code so that we check for different
blocks _before_ checking for different subtrees. There's no point to
keep matching once you're done with a block.

Differential Revision: https://phabricator.services.mozilla.com/D89365

UltraBlame original commit: b0674d532cabae27073e02b87f7622b15ae06f8f