Bug 1661720 - Don't claim to support monochrome for known PDF printers. r=jwatt
They don't. Right now we get to GetDataFromPrinter from
nsDeviceContextSpecWin with mPrintInColor = false, but this copy back
and forth ends up setting it to true:

  https://searchfox.org/mozilla-central/rev/2fb77a351ab34e9994f01377e2e12486a50f1737/widget/windows/nsDeviceContextSpecWin.cpp#448

Even if you avoid that, the printed file ends up being in color. This is
a bit of a hack, the right long-term fix is to detect whether the
printer supports monochrome in Gecko (add an attribute to nsIPrinter)
and communicate it to the front-end.

There's no equivalent to DC_COLORDEVICE but for monochrome in:

  https://docs.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-devicecapabilitiesa

I think we could check if DC_FIELDS returns the COLOR field or such.
I'll poke at doing that but this is less risky.

Differential Revision: https://phabricator.services.mozilla.com/D89306

UltraBlame original commit: d979307c639f0b494e7da2a85b0580db0c2cca6f