No bug: block update verify on upload-symbols. r=releng-reviewers,aki DONTBUILD
This fixes a bug that was introduced in recent optimizations where
upload-symbols may not run for shipped release builds. We can add it as
a dependency for this, or some other per-platform task that runs for
releases. As long as it runs before we ship this should be fine.

Differential Revision: https://phabricator.services.mozilla.com/D89494

UltraBlame original commit: 0f8808c2adf83c285c793c6ed553fc941818aa43