Bug 1663819 - Loosen the styling restrictions on meter and input[type=range] pseudo-elements. r=jwatt
These were there to prevent crashes because when the type of the frame
changes dynamically we would mess up and fail to find the right
insertion point.

We deal with this correctly nowadays:

  https://searchfox.org/mozilla-central/rev/b2716c233e9b4398fc5923cbe150e7f83c7c6c5b/layout/base/nsCSSFrameConstructor.cpp#8521-8536

And the list was never exhaustive in the first place (e.g., writing-mode
can also cause us to reframe, and so do a bunch of other properties). So
there's no real reason to be so strict anymore.

Also change ::-moz-complex-control-wrapper while at it too, because it's
not web exposed so there's no need for those declarations in the first
place.

Differential Revision: https://phabricator.services.mozilla.com/D89586

UltraBlame original commit: b0587b321c5b06a225a082ed8d5eb4e31e62e0e9