Bug 1663857 - Introduce the Pioneer enrollment ping. r=rhelmer
This introduces the enrollment ping (really the existing 'ping-study'
ping with a different schema name) that is sent both when user joins
Pioneer and when they join a new study.

The one sent when joining pioneer has a different namespace, because
we want this data to flow in a separate dataset. The one sent when
joining the study behaves like the deletion-request, has an empty
payload and is properly routed to the study table due to the rules
on the pioneer ingestion pipeline.

This additionally lands documentation required for data-review.

Differential Revision: https://phabricator.services.mozilla.com/D89718

UltraBlame original commit: 8872cde42a12a1c551ddaafa02734707cd300547