Bug 1663993: Add warm-up data field to SelfHostedLazyScript r=tcampbell
The issue here is that we trial-inline a self-hosted call, trigger relazification, then try to call the trial-inlined function. Unlike other functions, self-hosted lazy functions do not have a BaseScript. Instead, we have a fake `SelfHostedLazyScript` per-runtime that contains a trampoline pointer. In general, this is good enough for jitcode. However, when we call a trial-inlined function, we have to guard that it has a `BaselineScript`, and the first step is to check if it has a `JitScript`.

To make `branchIfScriptHasNoJitScript` work for self-hosted lazy functions, this patch adds a warm-up data field to `SelfHostedLazyScript`.

Differential Revision: https://phabricator.services.mozilla.com/D89848

UltraBlame original commit: b32cb06ef2aeaa939af0484498fcbfd74db05790