Bug 1664408 - Don't delete `~/.mozbuild/mozboot` directory after use in `bootstrap` r=mhentges,nalexander
The idea here was that we keep the NDK's downloaded in this directory as a "cache", such that if the download gets interrupted, then we can resume the download from an earlier point. This logic was walked back by bug 1637379, which deletes the cache.

In the spirit of the original version of the code, remove the `os.rmdir()` that was introduced by that bug.

An alternative fix for this is to download the NDK to a temporary directory and just clean that up entirely after every call to `mach bootstrap`, but then we would be forgoing the build cache behavior, which I'm not sure makes sense at this point in time.

Differential Revision: https://phabricator.services.mozilla.com/D90125

UltraBlame original commit: 7989a40e6c891daef7ff7ff97fedf5c0bf4a57a8