Bug 1664419. Need to suppress scrollbar repaint when seting visual viewport offset. r=kats,botond
In ScrollFrameHelper::ScrollToImpl we suppress repainting scrollbars when calling UpdateScrollbarPosition if we've decided not to schedule a paint (because we are sending a pending scroll update instead). Otherwise we repaint which is what we are trying to avoid.

The desktop zooming scrollbars code added a call to UpdateScrollbarPosition in SetVisualViewportOffset because the vv offset is what determines the scrollbar position, no longer the layout scroll position. This means that we can trigger a repaint via that path, and there is a test that checks that we don't.

This patch moves the SetVisualViewportOffset call to just after the code that determines if we need to schedule a paint (it was just before it), and then add the AutoScrollbarRepaintSuppression class.

Differential Revision: https://phabricator.services.mozilla.com/D89902

UltraBlame original commit: 9dd986a71b7171ce2b27b6c8c27e4b3e4c6dcdc6