Bug 1642735 - Suppress exceptions in ChromiumCDMProxy if we don't have MediaKeys. r=jbauman
When rejecting promises in ChromiumCDMProxy we pass an exception to MediaKeys to
represent the error that took place. However, if we do not have a any keys the
exception is not passed anywhere. Since these exceptions will assert on
destruction that they were handled we need to explicitly suppress the exception
when we don't have MediaKeys to avoid firing asserts.

The case we hit this issue in is during browser shutdown, so I think it makes
sense to ignore the exception. This is not a case of simply ignoring an
exception when it can be handled, this is that we're in a state where various
machinery is becoming unavailable and where it makes sense to not try and send
the exception any further.

Differential Revision: https://phabricator.services.mozilla.com/D90156

UltraBlame original commit: 5f67698e1e529218fa1d61b8f00340b9450d31f6