Bug 1664567 - Set `MOZ_OBJDIR` in builder scripts for CI r=glandium
Without setting this variable, the call to `mach python` in `taskcluster/scripts/builder/build-linux.sh` will create a `virtualenv` in the default location -- on Linux in CI, that's `obj-x86_64-pc-linux-gnu`. Everything else about the build is going to end up in `obj-build`, so just set `MOZ_OBJDIR` appropriately to point to that directory.

Doing so enables us to land bug 1663755, as it causes the logic added to `symbols_archive.py` as part of bug 1654994 to behave as expected.

Differential Revision: https://phabricator.services.mozilla.com/D89975

UltraBlame original commit: 025de03c734c81c91ab6981905ef555dfe4747d1