Bug 1651569 reduce required SNR to accept KissFFT ConvolverNode implementation r=padenot
as used in Firefox on some platforms.

minRequiredSNR = 83 permits the quadratic mean of errors to be up to
(10^8.3)^-0.5 ≅ 0.036 * 16384 * 2^-23, implying that most of the samples would
have error significantly less than one ULP for the representation of the peak
of the ramp (16384).

Differential Revision: https://phabricator.services.mozilla.com/D90209

UltraBlame original commit: 0194a0d1befc4a1afb4bc4d0cad36ca2b965d994