Bug 1663865 - Fix incorrect assertion check. r=nical
The debug assertion check here is trying to validate that we don't
enter an unexpected path for non-tiled images if the image template
has tiling enabled.

However, it was not correctly handling the case where an image key
has been sent but the image was removed. This case is handled by
the `resolve_image` code below, so it's a valid state to encounter
in this code path.

Differential Revision: https://phabricator.services.mozilla.com/D90343

UltraBlame original commit: 4107acc2e89b5125daee8f41e4ccb811ea8707b4