Bug 1654836 - Add a test for checkerboarding with zero displayport multipliers. r=nical
This ensures that even with the displayport margins dropped to zero we don't
get any checkerboarding. This currently fails with WR enabled, because the
displayport alignment code moves the displayport so it doesn't enclose the
unaligned displayport, and that's bad.

Differential Revision: https://phabricator.services.mozilla.com/D90501

UltraBlame original commit: 8b4f9898cdce9ccdb83d45424e51d8986708ea3e