Bug 1661513: Fix reporting of expected failures in fail-if tests. r=ahal
Without this change, failures in tests marked with `fail-if` are reported as
passes, though with messages and stacks that describe the failure. That works
as far as the automation machinery is concerned, but it is extremely confusing
to people who are trying to reproduce and fix the failures locally.

This change fixes the reporting so that they show up as expected failures in
logs, but have the same effect on final test status.

Differential Revision: https://phabricator.services.mozilla.com/D90616

UltraBlame original commit: 3b0e057ff3c395a2c0e8f48083e359fb77f49c72