Bug 1659516 - Remove compatibility.css import in tooltip.css that leads to the css being loaded twice r=mtigley,daisuke
If we open the Compatibility panel, the rules panel ill import the
tooltip.css for rendering the tooltips. However, the tooptip.css
imports compatibility.css and now the styles are not only unnecessarily
parsed but also on opening the Compatiblity panel, this stylesheet is
loaded and parsed again leading to the stylesheet
being loaded twice and hence these styles are applied twice for
each element in the compatibility panel unnecessarily.
We get all the necessary styles for the compatibility tooltip into
the tooltip.css (adding the necessary changes needed to handle the
color differences in the background) and remove the import of the
external compatibility.css

Differential Revision: https://phabricator.services.mozilla.com/D90533

UltraBlame original commit: 044db1a1d38daf6bf9890da4fd9c489d968b3cae