Bug 1664510 - Add localization for the WDBA r=flod,nalexander
This patch adds localization to the WDBA in a somewhat unusual manner. It's very difficult to localize toast notifications because the button text tends to overflow, which the localizers won't realize. Therefore, we are making a simpler version of the toast notification that only uses "Yes" and "No" as the button text. In English, we will not use this version and will continue using the version we have been using.

This means that the localized toast notifications will not have an "Ask me later" button, so the behavior will differ a bit from the English version. Specifically, there is no "followup" for the localized notification since there is no button that ought to cause it to be shown.

We've considered and rejected using localizer comments to keep these button strings short. We are limited by the amount of space the text takes up, not the number of characters. We just don't have a good way to convey this information to localizers in an actionable way.

Differential Revision: https://phabricator.services.mozilla.com/D90608

UltraBlame original commit: d6c9e691db15a4ed7388d0338f29b8af8907b32d