Bug 1659227 - Handle SVG{Inner,Outer}Frame as the clip frame for partial pre-render. r=heycam,botond
In SVG frame tree SVG{Inner,Outer}Frame is normally not scrollable but clips
descendants, thus nsLayoutUtils::GetNearestScrollableFrame is not able to find
the clip/scroll frame.  So in this change, we introduce
GetNearestOverflowClipFrame which is a similar function to
GetNearestScrollableFrame but it also checks whether the frame is SVGInnerFrame
or SVGOuterFrame which is clipping the given nsIFrame during walking up the
frame tree.

In this change we also change the logic to get the clip/scroll frame and the scroll
id as follows;

1) Get the nearest overflow clip frame in the same document
2) Use the root frame if there is no clip frame
3) Then try to find the scroll id if the clip frame is asynchronously scrollable

so that it becomes clearer before.  It's still messy though.

Differential Revision: https://phabricator.services.mozilla.com/D90647

UltraBlame original commit: 9bcb053caba37dd041391dbeea2020f0638956df