Bug 1667258 - Disallow JS::EncodeScript on run-once scripts. r=jandem,nbp
Use the same conditions for EncodeScript as we do for CopyScriptImpl. This
involves checking if the top-level script was compiled as run-once. This
check applies regardless of whether the script has been run on not (which is
a stricter test than might otherwise be required). This lets us avoid the
"cloneSingletons" mechanism altogether.

This is more consistent with browser use-cases of EncodeScript which do not
mark scripts as run-once.

We change test behavior to not use the run-once flag for non-incremental
mode. Cases that are testing run-once-only features now use incremental XDR
which is consisten with what the browser actually does.

Depends on D91363

Differential Revision: https://phabricator.services.mozilla.com/D91364

UltraBlame original commit: a12dfabdb0408b38573577b1d13bcafa30e4a79c