Bug 1667478. Fix dom/events/test/pointerevents/test_bug1303704.html to wait for scroll events properly. r=kats
This test is a little bit different but the same general principle applies as in the test fixes in

https://hg.mozilla.org/integration/autoland/rev/30cd33572b3d
and the followup fix to that
https://hg.mozilla.org/mozilla-central/rev/f51b6c3688a4

We also disable smooth scroll (all of the other previous tests fixed already had smooth scroll disabled) so we get one scroll instead of many small ones over time.

I didn't check that both of the two fixes are necessary or not.

Differential Revision: https://phabricator.services.mozilla.com/D91481

UltraBlame original commit: 7d6d00b58029ffb7048e947ad4c42af884837841