Bug 1667510 - Make nsDateTimeControlFrame not do math on nscoord_MAX. r=dholbert
This is not needed to fix the bug, but just makes the code a bit less
prone to this stuff.

Right now we rely on the child being a flex container and having a
concrete BlockStartAscent(). When that breaks, we do math with
nscoord_MAX (ASK_FOR_BASELINE), which causes all sorts of weird stuff to
happen.

Sanity-check the value code. If the child has no concrete baseline,
keeping the ASK_FOR_BASELINE in our own ReflowOutput will end up
synthesizing one as normal from the child, which is ok.

Differential Revision: https://phabricator.services.mozilla.com/D91620

UltraBlame original commit: dd1a9d5f7ae00f9d1310a7664cf496de78edfa3e