Bug 1667598 [wpt PR 25804] - The adjusted_start must be computed based on the part_start, a=testonly
Automatic update from web-platform-tests
The adjusted_start must be computed based on the part_start

In order to create a new ShapeResultView, we use segments of a previous
instance, analyzing its runs based on a new index range for the new
instance. Since the start/end of the RunInfo structure may be different
than the start index of the ShapeResultView, we use an adjusted_start
to adapt the logic to the different data structures.

We assume that if the passed start_index for the new range is greater
than the current run's part start being processed, we should define the
new run's start as the difference of these two values. However, we must
also ensure that start_index is at least equal to the the part_start
value for such run.

The problem described in the issue 1130153 comes from the fact that
we use the start_index on cases where it's smaller than the run's
part_start. This leads to an incorrect (longer than expected) run.

Bug: 1130153
Change-Id: Id5dd505baf1f990d2fe7a9406e191979cbd6b9f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2431523
Reviewed-by: Koji Ishii <kojiichromium.org>
Commit-Queue: Koji Ishii <kojiichromium.org>
Cr-Commit-Position: refs/heads/master{#811105}

--

wpt-commits: f2cb066328f2035c20b2a9726febd82056b7ac41
wpt-pr: 25804

UltraBlame original commit: 313bd552525f0010b423b7cd307c520dcfa8454a