Bug 1667116 - Request to stay unredirected on WR/OGL, too. r=stransky CLOSED TREE
Until bug 1617002 Firefox would unintendedly not get unredirected on X11
when used with WR and OpenGL as it didn't fullfill the requirements -
not being transparent.

Now that this is fixed, fullscreen unredirection can happen on certain
compositors. For the basic compositor we already explicitely ask to not
get unredirected - do that for all backends.

Differential Revision: https://phabricator.services.mozilla.com/D92180

UltraBlame original commit: 442928b108aa131d1aed5beca2516fc02d1356a4