Bug 1668458 - Load specialPowers web extension in gecko-only wpt tests, r=jmaher,twisniewski,whimboo
Lack of support for specialPowers is often cited as a reason for
developers to write mochitests rather than wpt. Although using
specialPowers means that the tests can't be upstreamed, having it
available in gecko-only tests should reduce the need to write
mochitests and potentially means that a patch which previously would
be written with just mochitests would instead be written with just
wpt, a fraction of which would be unsuitable for sharing.

Nevertheless a gecko-only specialPowers-using test should be
considered the option of last resort when cross-browser APIs such as
testdriver aren't sufficient.

This patch installs the specialPowers extension when the path to the
XPI is provided via a --specialpowers-path command line argument. To
ensure this only happens for gecko-only tests, the url_base is added
to the Test object, and the firefox settings() method is updated to
contain a specialpowers setting, which uses the url_base of the test
to determine if this is a gecko-only test.

Because the extension has to be enabled using marionette, we also need
to route this setting through the to executor, along with the provided
XPI path.

Differential Revision: https://phabricator.services.mozilla.com/D92035

UltraBlame original commit: 7be538be385a10b7e17a0b7cbcbe28c57e18e462