Bug 1651725 - Add .type property to WASM Memory, Table & Global r=lth
This is part of adding type reflections to the WebAssembly JS-API. This
adds a new property on the WebAssembly.Memory, WebAssembly.Table &
WebAssembly.Global objects. It exposes the current type or the object in
a format which is compatible with the constructor for type's
constructor.

Differential Revision: https://phabricator.services.mozilla.com/D90702

UltraBlame original commit: dff33d0b33c8fb168966504f5562572ddec72a8e