Bug 1669731 - P3. Add Allow API to ChannelClassifierService. r=timhuang,nhnt11
UI needs to distinguish the cases when a channel is shimmed and is
unshimmed.
When a channel is unshimmed, we unblock the channel and simply treat the channel as a
non-tracking channel.
When a channel is shimmed, although the channel is unblocked by
URLCLassifier, we still want to show it in the UI. For this case,
URLClassifier will notify a content blocking event when a channel is
unblocked.

This patch adds a new allow API, so the caller can use unblock() or
allow() depending upon the case it requires.

Differential Revision: https://phabricator.services.mozilla.com/D93271

UltraBlame original commit: 9754af354ad7ff31c9ac2c2b356d88f04f2557f5