Bug 1670643 - Ignore DOMWindowClose events fired at the preview browser. r=jfkthame
See the comment.

https://phabricator.services.mozilla.com/D90808 made it so that we
reused the print document (and thus <browser>) to do the actual print
job.

When a print job finishes, we call window.close(), which is used as a
signal to the front-end to clean up the <browser> element. Before that
patch we sent that signal to the browser created in printUtils.js, but
afterwards we send it to the preview window which causes this badness
when the event bubbles up.

Ignoring it effectively restores the previous behavior, so let's do it.

Differential Revision: https://phabricator.services.mozilla.com/D93499

UltraBlame original commit: 44fea8486cb1cbd21a2ca8b167007b2f5c931cf5