Bug 1654477 P1: Use a unique ID to track recorded textures. r=jrmuizel
This creates a unique (per content process) ID for each TextureRecorded. This is
then used as a lookup in the GPU process for the real texture.
Previously the pointer for the TextureRecorded's DrawTarget was used, but as the
translated texture is used in multiple threads in the GPU process, pointer reuse
could mean that the handle for an old texture could be picked up by mistake.

Differential Revision: https://phabricator.services.mozilla.com/D93501

UltraBlame original commit: f1c18620c5616bc606f56ee60e833d427847d165