Bug 1647313 - Fix update tests to wait on the update badge before they test for it r=mhowell
Updated tests to wait a moment for the update badge. The badge is added on a timeout, so it may not be there when we first look.

Some tests actually already use this mechanism:
  browser_doorhanger_bc_downloadOptIn_bgWin.js
  browser_doorhanger_bc_downloadAutoFailures_bgWin.js
I'm not sure why it wasn't used consistently, but it seems to fix the intermittent failures with browser_aboutDialog_fc_downloadAuto.js to add the mechanism.

Differential Revision: https://phabricator.services.mozilla.com/D93837

UltraBlame original commit: d71b5c02b502b6bf1376b72ca94928bfd98856c2