Bug 1670846: Treat line break characters in pre-formatted text as a11y paragraph boundaries. r=MarcoZ
In pre-formatted text (including textareas), text leaf nodes can contain line break characters ("\n").
Therefore, when searching for a paragraph boundary, we must match these.
Pivot can't return an ofset, only an Accessible, so we store the offset for the last match on the rule.
Because our origin point might be within a text leaf which contains a line break before/after it, we must also pass in the offset in our origin in this case so we can begin the text search from the right place.

Differential Revision: https://phabricator.services.mozilla.com/D93719

UltraBlame original commit: 1929ad3775f905dc8cc7bd567ef79211b3991fe0