Bug 1671284 - Don't drop APZ requests that have changed the zoom but not the scroll. r=botond
The APZCCallbackHelper code would only enter the codepath to set the presShell
resolution if the repaint request had a scroll position update. This seems silly
because really we care about whether or not the async zoom changed. The included
test exposes this silliness by demonstrating how the presShell resolution can
get stuck at an incorrect value because the necessary repaint requests get
ignored.

The patch now allows the SetResolutionAndScaleTo codepath to be entered if
there is an async zoom on the repaint request, so that we make sure to update
the presShell resolution even if the scroll position hasn't changed.

Differential Revision: https://phabricator.services.mozilla.com/D93687

UltraBlame original commit: 1171c5ea3c825d139a8bda3f938ca7b397503f60