Bug 1668892: Disable operator-dictionary-spacing-004.html Operator dictionary chunk 1 - lspace/rspace on Android; r=bc
This test is mysteriously failing under a try push with e10s-multi enabled on
Android. Since MathML is not super-important to us, I'd like to disable it on
Android and file a follow-up bug in MathML for the failure.

Differential Revision: https://phabricator.services.mozilla.com/D94151

UltraBlame original commit: eadc5e3e54e5ebbc871ed2748c7973b4557f1748