Bug 1662693 - Ignore no-store for root favicons. r=mossop
Usually we don't store icons returning a no-store header in response, because we
assume the website may be using them to indicate a state (with some kind of
notification badge).
Root favicons may be considered an exception though, since they are unlikely
to be replaced on-the-fly to show badges, and we use them to guess a favicon
when a page doesn't specify one.

Differential Revision: https://phabricator.services.mozilla.com/D93802

UltraBlame original commit: 9822782abba0e4a24dbd35b6538c0afa5a5839fe