Bug 1670541 - Return the parent line end offset if an embedded character  contains a line break, r=Jamie
If the offset for an embedded character is queried, and it is an inline element, such as a link, and it contains a line break, return the full parent line.

Likewise, if querying character offsets before that embedded character, include its line break in the end offset.

Differential Revision: https://phabricator.services.mozilla.com/D94024

UltraBlame original commit: 5893300d26cb499de9480ae59e14c5981ca71944