Bug 1641595 - inline constants, part 2: random other cases. r=jseward
Instead of loading instruction-internal simd constants separately (on
x86/x64) and thus potentially increasing register pressure by
requiring temps, load them as part of the operation.

This patch handles a few remaining single-use-of-loaded-const cases,
none of which are probably all that important, this is just hygiene.

In the case where a loaded const is used multiple times, or a single
use can't be gotten rid of without introducing additional moves, we
assume (for now) that it's OK to load the constant separately.  The
cases that remain in this class are blends (that will use a mask and
its negation) and a few floating-point slow paths (some with multiple
uses, some with knotty logic).

Differential Revision: https://phabricator.services.mozilla.com/D94557

UltraBlame original commit: b5f6faabcb092fe073603a6d832cfbe009dc7ff8