Bug 1640662 - wasm ion simd register preferences part 1: unary-ish ops. r=jseward
Use useRegister / useRegisterAtStart / define / defineReuseInput (as
well as comments) so as to generate (and document) the best code under
context-less assumptions.  See comment 0 on the bug for a more general
description of how we use those functions.

We add test cases to assert that the best code is actually generated
when that makes sense.  There are test cases only for operations where
the operation itself has explicit, but optional, reg->reg moves at the
beginning of the emitted code to deal with register misallocation.

The test cases assert that extraneous moves are not inserted at the
start of an operation under ideal circumstances, taking into account
whether the operation prefers src == dest or src != dest.  For cases
where the operation prefers src == dest the input register is set up
to be the same as the output register, ie, we use local 0 as the
argument and generate a result in the result register; for cases where
the operation prefers src != dest, we use local 1 as the argument.

The test cases are fairly delicate - they may be sensitive to small
changes in Ion; comments in the test files highlight this, and the
tests are not meant to constrain Ion, only to highlight accidental
changes.  That said, these tests have already uncovered bugs, and for
now let's see if we can make them stick.

Drive-by fixes:

 - prefer AND-with-literal-mask and XOR-with-literal-mask in a couple
   of places (instead of preloading the constant) to generate better
   code, mostly where this matters right now because we introduce
   tests that need to inspect the code; a more general cleanup of this
   type of pattern may happen later

 - remove unused/unimplemented generality in the negFloatNNxM,
   absFloatNNxM, and notTypeNNxM primitives

 - make the reuseInput*() primitives assert that their inputs are
   SIMD registers, and adapt client code accordingly; this
   prevents some bugs since float registers do not compare equal
   if they are not of the same type

 - make moveSimd128{Int,Float} no longer emit a move if the source and
   destination are the same register

Differential Revision: https://phabricator.services.mozilla.com/D93630

UltraBlame original commit: 480aa6fe2e0a09be7f4003a1b776f9ab6e525ad6