Bug 1671998 - Clean up various move operations.  r=jseward
This is a cleanup patch.

Remove all 'src != dest' guards on moveSimd128, moveSimd128Float, and
moveSimd128Int, because they all have that guard built in.

Rewrite some moves that are under such guards to use that family of
masm instructions, and remove the guard.

In a few cases where we continue to use movaps, movapd, and movdqa
directly, make sure we use the right one for the mode (int or float)
we're currently in.  This required adapting one whitebox test.

Differential Revision: https://phabricator.services.mozilla.com/D94074

UltraBlame original commit: 72178fc22f42f264ae8bcdb6b241e258c33799c9