Bug 1657246 - Unset AudioCallbackDriver::mStarted on error. r=padenot
Start() will assert !mStarted, but only Stop() unsets it. In case we receive an
error through the state callback and try to restart the driver we fail the above
assert. This patch unsets mStarted on error, and moves the Stop() unsetting to
the STOPPED part of the state callback, to contain all the unsetting there.

Differential Revision: https://phabricator.services.mozilla.com/D94564

UltraBlame original commit: e7d0acdef9c3f01bf55c1eb91b7a431203f5a241