Bug 1669861 - Use the visual scroll offset consistently for DisplayPortMargins computations. r=kats
Even when the margins don't come from APZ, they are relative to
the visual viewport, and for the RCD-RSF there may be an offset
between the visual and layout viewports that they should be
translated by.

Bug 1669982 modified the call site in
CalculateAndSetDisplayPortMargins() to do this. This patch
refactors things such that all setters of the display port
do this if appropriate.

Differential Revision: https://phabricator.services.mozilla.com/D93428

UltraBlame original commit: 93021c9bf18c4f282e985d44cb0fd660bc653e6b