Bug 1673992 - [taskgraph] Ensure we stop extracting mozharness.zip in the cwd r=bhearsum
Rather than putting a top-level "mozharness" directory in "mozharness.zip",
this puts the root of mozharness at the root of the archive. Then we extract
the archive into a "mozharness" directory. This way the path on disk ends up
being the same, but generic-worker won't attempt to modify the permissions of
files in the cwd.

Differential Revision: https://phabricator.services.mozilla.com/D95138

UltraBlame original commit: e281605f0fd24d3808d4af5964213588ef7c145a