Bug 1671460 - Update to apilint 0.4.0. r=snorp
There is an unfortunate-but-necessary ordering issue that needs to be fixed in
the api.txt file which is causing a lot of one time updates. The ordering of
annotations was not stable (and now it is).

Differential Revision: https://phabricator.services.mozilla.com/D95382

UltraBlame original commit: 73ef4114f8c1d015bdb331c8dfb1d0d052cd8b98